app/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
djangorestframework==3.14.0
kombu==5.4.2
psycopg2-binary==2.9.9
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
pytest==7.4.3
django_celery_beat==2.7.0
celery[redis]==5.4.0
langchain==0.3.15
langchain-community==0.3.15
langchain-openai==0.3.1
cohere==5.13.11
twilio==9.4.3
langgraph==0.2.69